+++ RP_Notify/Program.cs
@@ -8,7 +8,7 @@ using RP_Notify.PlayerWatcher.MusicBee.API;
 using RP_Notify.PlayerWatchers.Foobar2000.BeefWebApiClient;
 using RP_Notify.RpApi;
 using RP_Notify.SongInfoListener;
-using RP_Notify.Toast;
+using RP_Notify.ToastHandler;
 using RP_Notify.TrayIcon;
 using System;
 using System.Windows.Forms;
@@ -26,7 +26,7 @@ namespace RP_Notify
                 .AddSingleton<ILog, Log>()
                 .AddSingleton<RestClient>()
                 .AddSingleton<IRpApiHandler, RpApiHandler>()
-                .AddScoped<IToastHandler, ToastHandler>()
+                .AddScoped<IToastHandler, ToastHandler.ToastHandler>()
                 .AddSingleton<Foobar2000Watcher>()
                 .AddSingleton<MusicBeeIPC>()
                 .AddSingleton<MusicBeeWatcher>()
